sql/requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
certifi==2024.2.2
charset-normalizer==3.3.2
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docopt==0.6.2
executing==2.0.1
Faker==24.4.0
fastjsonschema==2.19.1
filelock==3.13.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
ipython==8.12.3
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==8.10.0
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.26.0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
PyYAML==5.4.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
systemd-python==234
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
virtualenv==20.25.1
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==1.0.0
